feat: Update Solana Actions route for blinks to include cluster selection
The code changes in this commit modify the Solana Actions route for blinks to include the ability to select a cluster. Previously, the cluster was hardcoded to "devnet", but now it can be dynamically set based on the value of the environment variable NEXT_PUBLIC_WALLET_ENV. This allows for greater flexibility in connecting to different Solana clusters.