feat: Update dashboard page with menu for creating different types of items
The code changes in this commit update the dashboard page to include a menu for creating different types of items. Previously, there was a single "Create" button, but now it has been replaced with a menu that allows the user to select from multiple options such as fundraisers, GitHub projects, and DRiP items. This provides a more intuitive and flexible user experience for creating new items on the dashboard.